pyproject.toml
[project]
name = "auth-sprint-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "async-fastapi-jwt-auth>=0.6.6",
    "fastapi-pagination>=0.12.34",
    "fastapi[standard]==0.115.4",
    "orjson>=3.10.15",
    "passlib>=1.7.4",
    "psycopg>=3.2.4",
    "psycopg-binary>=3.2.4",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-mock>=3.14.0",
    "python-jose[cryptography]>=3.4.0",
    "redis==5.2.0",
    "setuptools>=75.8.0",
    "sqlalchemy>=2.0.38",
    "typer>=0.15.1",
    "uvicorn>=0.34.0",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "flake8>=7.1.2",
    "isort>=6.0.1",
]
